Updated to latest Core version; Wallet and Node are now in sync

+++ NodeWebApi/Dtos/Transactions/TransactionDto.cs
@@ -9,7 +9,7 @@ namespace NodeWebApi.Dtos.Transactions
         public DateTimeOffset CreationTime { get; init; }
         public byte[] MerkleHash { get; init; }
         public byte[] Input { get; init; }
-        public int Amount { get; init; }
+        public long Amount { get; init; }
         public byte[] Output { get; init;}
         public bool IsDelegating { get; init; }
         public byte[] Signature { get; init; }
